Refactor UserManager and update menu options
UserManager instantiation was moved to specific methods for application efficiency, and its menu options were updated for better understanding and functionality. Additionally, unnecessary import statements were removed, and new utility function for adding a new line was created for improved code readability. A method for adding a book was also added to the BookManager.