Refactor LibraryMember and UserManager interactions
The LibraryMember object creation has been refactored to remove the hard-coded dummy data. A helper method has been created to collect new member details. Additionally, UserManager's usage of the UserTableManager has been simplified, changing the way a new library member is added to further separate responsibilities. The UserManager now performs the check if a member exists in the local list before trying to add a new member to the database via the UserTableManager.